Fixed some more warnings (god there are so many)

+++ xash3d_engine/engine/ref/gl/gl_rlight.c
@@ -66,31 +66,31 @@ void CL_RunLightStyles(void)
 
 		if ( !ls->length )
 		{
-			tr.lightstylevalue[i] = 256 * scale;
+			tr.lightstylevalue[i] = (int)(256.0f * scale);
 			continue;
 		}
 		else if ( ls->length == 1 )
 		{
 			// single length style so don't bother interpolating
-			tr.lightstylevalue[i] = ls->map[0] * 22 * scale;
+			tr.lightstylevalue[i] = (int)(ls->map[0] * 22 * scale);
 			continue;
 		}
 		else if ( !ls->interp || !CVAR_TO_BOOL(cl_lightstyle_lerping) )
 		{
-			tr.lightstylevalue[i] = ls->map[flight % ls->length] * 22 * scale;
+			tr.lightstylevalue[i] = (int)(ls->map[flight % ls->length] * 22 * scale);
 			continue;
 		}
 
 		// interpolate animating light
 		// frame just gone
-		k = ls->map[flight % ls->length];
+		k = (int)ls->map[flight % ls->length];
 		l = (float)(k * 22.0f) * backlerp;
 
 		// upcoming frame
-		k = ls->map[clight % ls->length];
+		k = (int)ls->map[clight % ls->length];
 		l += (float)(k * 22.0f) * lerpfrac;
 
-		tr.lightstylevalue[i] = (int)l * scale;
+		tr.lightstylevalue[i] = (int)((int)l * scale);
 	}
 }
