+++ xash3d_engine/engine/ref/soft/r_light.c
@@ -69,31 +69,31 @@ void GAME_EXPORT CL_RunLightStyles(void)
 
 		if ( !ls->length )
 		{
-			tr.lightstylevalue[i] = 256 * scale;
+			tr.lightstylevalue[i] = (int)(256.0f * scale);
 			continue;
 		}
 		else if ( ls->length == 1 )
 		{
 			// single length style so don't bother interpolating
-			tr.lightstylevalue[i] = ls->map[0] * 22 * scale;
+			tr.lightstylevalue[i] = (int)(ls->map[0] * 22.0f * scale);
 			continue;
 		}
 		else if ( !ls->interp )  // || !CVAR_TO_BOOL( cl_lightstyle_lerping ))
 		{
-			tr.lightstylevalue[i] = ls->map[flight % ls->length] * 22 * scale;
+			tr.lightstylevalue[i] = (int)(ls->map[flight % ls->length] * 22.0f * scale);
 			continue;
 		}
 
 		// interpolate animating light
 		// frame just gone
-		k = ls->map[flight % ls->length];
+		k = (int)ls->map[flight % ls->length];
 		l = (float)(k * 22.0f) * backlerp;
 
 		// upcoming frame
-		k = ls->map[clight % ls->length];
+		k = (int)ls->map[clight % ls->length];
 		l += (float)(k * 22.0f) * lerpfrac;
 
-		tr.lightstylevalue[i] = (int)l * scale;
+		tr.lightstylevalue[i] = (int)((int)l * scale);
 	}
 }
 
@@ -334,9 +334,9 @@ static qboolean R_RecursiveLightPoint(
 
 			if ( tr.ignore_lightgamma )
 			{
-				cv->r += lm->r * scale * 2.5;  // scale;
-				cv->g += lm->g * scale * 2.5;  // scale;
-				cv->b += lm->b * scale * 2.5;  // scale;
+				cv->r += (unsigned int)((float)lm->r * (float)scale * 2.5f);  // scale;
+				cv->g += (unsigned int)((float)lm->g * (float)scale * 2.5f);  // scale;
+				cv->b += (unsigned int)((float)lm->b * (float)scale * 2.5f);  // scale;
 			}
 			else
 			{
